backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.7.2
Pillow==10.4.0
pydantic==2.9.0
